backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
httpx==0.25.2
python-jose==3.3.0
deepgram-sdk==2.11.0
openai==1.3.5
jira==3.5.1
supabase==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
bcrypt==4.0.1
aiohttp==3.8.6
python-jose==3.3.0
python-multipart==0.0.6
aiohttp==3.8.6
python-dotenv==1.0.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
openai==1.12.0 